Logica y funcionamiento para el crud de facturas
Completados los archivos para hacer funcional el crud de facturas

+++ TFG/src/main/java/com/example/TFG/repository/FacturaRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 
-public interface FacturaRepository extends JpaRepository<Factura, Long> {
+public interface FacturaRepository extends JpaRepository<Factura, Integer> {
 
     @Query("SELECT f FROM Factura f WHERE f.empresa.id_empresa = :idEmpresa")
     List<Factura> findFacturasByEmpresaId(@Param("idEmpresa") Long idEmpresa);
